MMW project update 1
The ones with the .proj are the landing page. However, if there is no proj within the filename, it is the website itself. This is for my first ever semi large scale project where I need to build a landing page and a front-end website to showcase our product. This also serves as a project for my portfolio for my proven knowledge and application of HTML, CSS, and JavaScript.